реструктуризация проекта + компонент создания ревью и компонент профиля. компонент просмотра ревью у контентов

+++ pkg/models/models.go
@@ -21,7 +21,7 @@ type Review struct {
 	ContentID int
 	UserID    int
 	Review    string
-	Date      int
+	Date      int64
 }
 
 type Content struct {
